refactor: swap saving input change from ls to redux store
