+++ rss-react-components/package.json
@@ -12,10 +12,13 @@
     "test": "vitest"
   },
   "dependencies": {
+    "@reduxjs/toolkit": "^1.9.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.43.9",
-    "react-router-dom": "^6.9.0"
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.9.0",
+    "redux": "^4.2.1"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.16.5",
