+++ rss-react-components/src/app/rootReducer.ts
@@ -1,0 +1,8 @@
+import { combineReducers } from '@reduxjs/toolkit';
+import storeInputReducer from '../components/Search/searchSlice';
+
+export const rootReducer = combineReducers({
+  search: storeInputReducer,
+  // searchResults: dataCardsReducer,
+  // formSubmissions: formCardsReducer
+});
