+++ rss-react-components/src/app/store.ts
@@ -1,0 +1,8 @@
+import { configureStore } from '@reduxjs/toolkit';
+import { rootReducer } from './rootReducer';
+
+export default configureStore({
+  reducer: {
+    rootReducer,
+  },
+});
